+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+### [106.6.8](https://github.com/Sage/carbon/compare/v106.6.7...v106.6.8) (2022-06-01)
+
+
+### Bug Fixes
+
+* **decimal:** ensure that value prop is reliably displayed ([5130b96](https://github.com/Sage/carbon/commit/5130b96e7f5f7cac18962c7c6f4ecd60f89e014c))
+
+### [106.6.7](https://github.com/Sage/carbon/compare/v106.6.6...v106.6.7) (2022-05-31)
+
+
+### Bug Fixes
+
+* **menu:** change passive menu item background color ([a05a3ad](https://github.com/Sage/carbon/commit/a05a3add1b854a1dca4a6fcf19c1df0ce78a3efb)), closes [#5064](https://github.com/Sage/carbon/issues/5064)
+
+### [106.6.6](https://github.com/Sage/carbon/compare/v106.6.5...v106.6.6) (2022-05-24)
+
+
+### Bug Fixes
+
+* **button:** stop icon from wrapping when nowrap prop set ([0711c3f](https://github.com/Sage/carbon/commit/0711c3fc57d45167ce51e2e64e1ac9cda7c950b4)), closes [#5101](https://github.com/Sage/carbon/issues/5101)
+
+### [106.6.5](https://github.com/Sage/carbon/compare/v106.6.4...v106.6.5) (2022-05-23)
+
+
+### Bug Fixes
+
+* **heading:** remove doc comments to fix type definitions ([9903e45](https://github.com/Sage/carbon/commit/9903e45586141eca5f7d1afc2de7349932d755e4))
+
 ### [106.6.4](https://github.com/Sage/carbon/compare/v106.6.3...v106.6.4) (2022-05-19)